+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added "Find by ID" helper dialog.
     - https://github.com/StateSmith/StateSmith-drawio-plugin/issues/33
+- Visibility fix for upstream vscode draw.io visibility issue.
+    - https://github.com/hediet/vscode-drawio/issues/382
 
 ---